Split code into main.rs and lib.rs
